Decrease number of machines in e2e tests
The e2e tests are currently very unstable. I suspect lack of resources
in the devstack to be the issue. In an attempt to solve this, the number
of control-plane nodes are reduced in this commit. Now only the multi-az
test will use 3 CP. The rest will have 1 CP only.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>