ImageFilter - add validations, switch to pointers
This commit changes `ID` and `Name` of `ImageFilter` to pointers which
should only affect go marshalling.

Other than that it adds CEL validation of the ImageFilter, so that Name
or Tags can only be set when ID is unset. Conversions are updated
accordingly to make sure we only set Name when ID is unset.

Moreover validation is added that ID has to be UUID. It's not enforced
in conversions, as non-UUID IDs would produce clusters or machines that
would not work properly.