Allow explicitly empty volume AZ
Replaces an AvailabilityZone string for volumes with a
VolumeAvailabilityZone struct which allows more flexibility in
defaulting behaviour. Specifically it enables us to express both the
current default behaviour where we take the volume AZ from the Machine,
and a new default behaviour where to don't specify a volume AZ at all.

In making this change to both RootVolume and AdditionalBlockDevices we
use common code for both APIs. This has the result of updating
RootVolume to be consistent with AdditionalBlockDevices.