instance: re-add error event if server creationg returned an error
When creating a server, Nova can return an error immediately after the
call was made to create it or later.

If it's returned immediately, we would like to restore the error event
that we used to have and that we actually test in our e2e when providing
a wrong availability zone.