Wait for Octavia-managed ports to be removed
In a best-effort mode, when cleaning a load-balancer, wait for the ports
with a device ID (mapped with the LB ID) and a certain prefix is deleted
(by Octavia itself, not CAPO managed) before claiming the LB is really
deleted.

This will avoid the reconcile to fail later when trying to remove the
network while some ports are still attached.