e2e: Fix deletion of test security group
We were trying to delete the security group in a DeferCleanup() called
from the test itself. This will execute before the cluster is deleted,
so it can't work as the cluster machines are still using the security
group. We need to delete the security group after the cluster has been
deleted.

In making this change, we also add a context argument to
postClusterCleanup and propagate that to the other users of this
mechanism.