e2e: Only dump resources on node1
We were calling DumpOpenStack* in the 'all nodes' argument to
SynchronizedBeforeSuite, but only running the checks in the 'node 1'
argument to SynchronizedAfterSuite. This is inefficient if we run with more
than 1 node, as only data dumped by node 1 will be checked.

This change moves the checks to a DeferCleanup() node called from the
creation node, so the cleanup will only run where it was created.

We also refactor CheckResourceCleanup to return error so we can return
its output directly to DeferCleanup.