Support Flavor IDs
If I have a system that uses flavor IDs, beacuse it's unambiguous, I
then need to translate it into a flavor name before passing it to CAPO
that will then translate it back into an ID using a list API call.
Adds a "FlavorID" field to the CRD, and handles resolution in the same
manner as images are handled.