Support Flavor IDs
If I have a system that uses flavor IDs, beacuse it's unambiguous, I
then need to translate it into a flavor name before passing it to CAPO
that will then translate it back into an ID using a list API call.
Check whether the "name" looks like a UUID, and use it directly for the
flavor lookup if you can.  Ideally, we'd make API changes like how
images are handled, where you can be explicit about it, but I guess all
OpenStack clients do this anyway!