Add ORC with image controller
This commits creates a new home for ORC, initially incubated in CAPO.
The initial implementation of ORC has an image controller.

The image controller is integrated with OpenStackServer and from there
into both OpenStackMachine and the bastion.

The E2E tests are updated to use the new image controller. Images are
now loaded dynamically by the test suite.