Re-allow `OpenStackMachine` to use their own `IdentityRef`
I think it's a regression when implementing the `OpenStackServer`
controller but we lost the ability to override the `IdentityRef` for the
`OpenStackMachine` (and therefore `OpenStackServer`).

This patch fixes that by using the `IdentityRef` from the
`OpenStackMachine` if it does exist, and fallback on the
`OpenStackCluster`'s `IdentityRef` which is required in v1beta1.