🐛 Fix support for `OpenStackMachine.Spec.SecurityGroups`
Somehow we broke that feature, probably when converting to
`OpenStackServer` and our CI missed that due to the lack of tests.

Now we add it back and we have tests.
This patch will add the security groups part of
`OpenStackMachine.Spec.SecurityGroups` in each Port that is in
`OpenStackServer.Spec.Ports`.