Ensure that existing ports also have correct tags and trunks
If port creation fails in the middle, and cleanup also fails,
then we may end up with ports with missing tags or trunks.
This could happen when hitting rate-limits for example or
if there is a network outage. This commit addresses the issue
by going through existing ports and ensuring that they have
correct tags and trunks.

Co-authored-by: Huy Mai <huy.mai@est.tech>
Signed-off-by: Lennart Jern <lennart.jern@est.tech>