Add Image templates and update dev docs
This adds a template for ORC managed Images. The template is meant as a
simple way for developers to get a bastion and node image loaded into
their openstack environment. Docs are also updated to reflect this.

The ClusterClass template is also updated to make it more useful. Most
notably, this adds variables for identityRef and for specifying the
image name. It is possible to choose between the traditional image including kubernetes components,
or using e.g. flatcar with a sysext overlay. The sysext overlay patch is only for ignition for now.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>