Add Image templates and update dev docs
This adds a template for ORC managed Images. The template is meant as a
simple way for developers to get a bastion and node image loaded into
their openstack environment. Docs are also updated to reflect this.

The ClusterClass template is also updated to make it more useful. Most
notably, this adds variables for identityRef and for specifying the
image name.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>