devstack: build OVN from source
In this PR we'll control the version of OVN to align with the upstream
devstack jobs building OVN and OVS from LTS versions.
This is largely a workaround to address some timeout issues that we
recently have and that with a controlled version of OVN we hope to
reduce.

(cherry picked from commit b79ec20333de776eeb8a64232954471d753b6158)