Stop printing InstanceState for OpenStackMachine
`InstanceState` was copied to the `OpenStackServer` API and the
`OpenStackMachine` controller doesn't care about this field anymore.
The field still exists for backward compatibility but it was decided
long time ago (when we worked on the `OpenStackServer` API) that it
wouldn't be populated anymore.
We missed to stop printing this field when listing the
`OpenStackMachines`.

Also, we now document that the field is not populated anymore. This was
also a miss.