Add ignore file for trivy
We have a couple of CVEs that require bumping the go version
to 1.23, which we do not want to do on the release branch.
Govulncheck also identified that we are not affected by the CVEs
even though we are on an affected version of the modules.
This can be verified by running govulncheck with -show verbose
and cross referencing the vulnerabilities between trivy and
govulncheck.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>