E2E: clusterctl-upgrade from v0.10
This is an attempt to fix the clusterctl-upgrade test that is currently
failing on this branch.
Starting from v0.9 is probably part of the issue since this is so old an
unsupported now.
We were also upgrading to CAPI v1.9.3, while the test module was at
v1.9.7. That could make the clusterctl-upgrade test incompatible with
the controllers potentially.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>