Exclude govulncheck from verify target
It was not intentional to make govulncheck run as part of
PR tests. The PR tests run the verify target. This commit
removes the govulncheck from that target.
The PR tests are anyway running with a different go version
than what we use to build release artifacts, so we cannot
really rely on the results.

Govulncheck will still be run in the weekly security-scan.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>