Improve handling of missing load balancer permissions
Currently, when a user tries to create a cluster using OpenStack
credentials which are missing the load balancer permissions, CAPO
adds the finalized to the OpenStackCluster resource then fails to
create the load balancer. When the user then tries to delete the
cluster, CAPO makes a GET request to the Octavia API to get the
load balancer details and receives a 403 (permission denied)
response, so the only way to allow the cluster deletion to
proceed is to manually remove the finalizer from the
OpenStackCluster resource.

This change prevents the above edge case by only attempting to
delete the API server load balancer if the load balancer ID is
populated in the OpenStackCluster's status field.