Set base for golanci-lint update PR
We cannot run this on PRs since that would require more privileges. We can still manually trigger it to run on a specific branch as a way of testing
the workflow.
Also fix the syntax for add-paths. It should be a comma or newline separated list of paths, but we had an extra "-" in front of the path.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>