Refactor generate-codegen
This is a manual backport.
This commit changes how we run and build codegen.
Instead of using a script inside the openapi-gen package, we now build and call the binary directly.
This avoids issues where the upstream package is incompatible with the go version we use.

Minor changes to generated code:
- api_violations.report: One additional violation detected by newer gengo
- zz_generated.openapi.go: Backslash escaping difference (cosmetic)

Signed-off-by: Lennart Jern <lennart.jern@est.tech>