Fix calcualtion of previous tag for release note generation
We used to generate release notes after tagging, so it was possible to just
list the tags and pick the one before the current. This is no longer the case.
Now we need to add the upcoming tag to the list since we generate the release
notes before creating the tag.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>